Move running to the constructor to reduce number of possible states.
This lets us remove the lazy-compute logic from score() (which is a user-provided method and so woul dhave to be repeated or factored out with additional complexity).

Move ruleset specification into the class; it generally needs to be specified twice from outside otherwise, and the rest of the class parametrization is specific to the ruleset anyway.

Move scoreParts initialization to its own method so users don't have to override the constructor, which is slightly more awkward than defining another method, thanks to the need to call super().